refactor: migrate appProvider to context folder
